Fix superclass for UNHANDLED- errors.
These errors represent unhandled entry types in a tar file, i.e. with
a particular type flag. The report functions for these conditions
expect to print a typeflag value, and there's an UNHANDLED-ERROR class
with a typeflag slot, but none of these conditions inherit from it, so
they fail to print when signalled. Fix the inheritance so printing
works.